Fix ecma_string_of_float assert.
The max length of an ECMAScript float string representation depends on the specifications [6.1.6.1.20](https://tc39.es/ecma262/multipage/ecmascript-data-types-and-values.html#sec-numeric-types-number-tostring), they are respectively[1] 22, 19, 25 and 24 for 6.a, 6.b, 6.c and 7. As we can see in the case 6.c they can be as long as 25, so the assert in the case of `ecma_string_of_float` is off by one.

Example of an ECMAScript float of length 25: -0.0000013475801431957922

[1] details:

6.a. minus + max_n = 1 + 21 = 22
6.b. minus + max_k + dot = 1 + 17 + 1 = 19
6.c. minus + zero + dot + -min_n + max_k = 1 + 1 + 1 + 5 + 17 = 25
7.   minus + max_k + dot + e + sign + exponent = 1 + 17 + 1 + 1 + 1 + 3 = 24